Added build info
- At connect time, build info is output in a notice with the form: 'Built: "2015-02-10T21:30:55+0000" from "git@github.com:Psiphon-Labs/psiphon-tunnel-core.git" at rev "d3a9bdcc04e1e33d7595b3ba06515785180eecd6"'
- Build info values are inserted into code via ldflags at build/link time.
- If values are not supplied, there is no problem -- the build values will just be empty.

The reason I chose to use the "-ldflags -X" trick rather than "go generate" is that it's more OS/environment agnostic. For example, the "date" command is very different between Linux and Windows, or git might not be present on the build machine (but the repo/rev could still be passed in).