Implement client-side tunnel duration reporting
* Client records the precise tunnel duration and
reports it in the next "status" request for the next
tunnel.
* To retain duration data, unreported tunnel durations
are persisted to the data store until reported.
* This new scheme supersedes the server-side duration
estimation based on pairing "connected" and final
"status" requests. The "connected" parameter in the
status request no longer follows the convention of that
old scheme.
* Rename "Session" to "ServerContext": a session is now
considered to be the entire lifetime of a Psiphon app
between user-commanded "Start" and "Stop". Renamed the
per-tunnel server state to avoid confusion.
* Heavily reworked serverApi code.
* Changed transferstats API, replacing json.Marshaler
interface with explicit ServerStats struct. Then moved
the "status" request request body formatting to serverApi
in the "psiphon" package.