Remove panic recover
We now believe that it's safer to allow panics to terminate the process and
rely on panic monitoring to address the underlying issue.

In the case of bugs in the Go runtime, random memory corruption is possible.
In past cases, this lead to "impossible" panics, such as an index out of
bounds immediately after a bounds check. As the program state is no longer
sound, it's better to terminate rather than continue on and risk arbitrary
execution; e.g., security checks may no longer be correctly enforced.

Another risk with panic recovery is failure to clean up resources. For
example, code below the recovery may have acquired a mutex lock and panicked
before releasing the lock.

In this recover case, the code below the recover used to be simpler, but now
makes calls back into TunnelServer and it's more difficult/risky to reason
about resource cleanup.