Emit ApplicationParameters as a single object
This allows the outer clients to more easily detect when an individual
ApplicationParameters value has been removed, as compared with the previous
scheme which would emit no notice for the removed field.

Duplicate ApplicationParameters notices are still skipped, but the entire
ApplicationParameters object is emitted if any field is changed, added, or
removed.