Fix: always log all fronting fields for remote_server_list
The meek_* fields in remote_server_list are only set with fronting values,
never direct dial addresses, and so the
requestParamLogOnlyForFrontedMeekOrConjure/etc. flags, which are intended to
avoid logging direct tunnel server addresses, do not apply, and result in the
unintended omission of flagged fields due to the lack of a relay_protocol
field.